JIT: move the numeric binop emission primitives into state/binop.rs
Step 2: a pure move of the register-allocated fixnum paths
(binop_integer*), the xmm float paths (binop_float, gen_cmp*_float), the
constant folds and the operand-load discipline (gp_ensure,
load_binary_(ret_)fpr) out of compile/binary_op.rs into the state module,
where the register-allocator internals they touch live. The inline
generators in builtins/numeric/ will drive the five pub(crate) entry
points (binop_integer / binop_float / gen_cmp_integer / gen_cmp_float /
gen_cmpbr_integer, plus the new gen_cmpbr_float and the fold probes);
everything else stays private to state. No behavior change.
